+++ lib/defaults.py
@@ -1,27 +1,27 @@
 #
-# PYro defaults configuration file
-#   Chris Martin (c) 2015
+# PYroMat defaults configuration file
+#   Chris Martin (c) 2015,2017
 #
-# When PYro loads config files, it automatically looks for 'defaults.py' 
+# When PYroMat loads config files, it automatically looks for 'defaults.py' 
 # in the installation directory.  Here, administrators can configure 
 # advanced options like allowing users to keep their own config files 
 # that override the defaults or maintaining their own registries and data
 # in their home directories.
 #
 # Configuration files are ordinary python scripts that are executed when 
-# PYro decides to load its configuration options.  PYro inspects the 
+# PYroMat decides to load its configuration options.  PYroMat inspects the 
 # locals() dictionary to see what (if any) variables were created.  If 
-# the parameters are recognized, they are loaded into the pyro.config 
+# the parameters are recognized, they are loaded into the pyromat.config 
 # dictionary.  This means that script files are just ordinary python code.
 #
-# PYro uses an unusual configuration loading scheme where no configuration 
+# PYroMat uses an unusual configuration loading scheme where no configuration 
 # parameter is ever truly overwritten.  Instead, values are appended to a
 # list of all values ever assigned to the parameter.  In this way, 
 # parameters that are intended to be configurable for multiple values 
 # need no special treatment, and it is easy to see a parameter's history.
 #
 # The line "myparam = 'value'" in a config file will result in
-#
+# >>> import pyromat as pyro
 # >>> pyro.config{'myparam'}
 # ['initial_value', 'value']
 #
@@ -41,18 +41,18 @@
 
 
 #** Configuration files **
-# This tells PYro where to find other configuration files. Keep in mind 
+# This tells PYroMat where to find other configuration files. Keep in mind 
 # that the last file loaded will be given precedence.
 #
 #> config_file = '/path/to/my/file'
 #
 #   or, perhaps
 #
-#> config_file = '/etc/pyro.py'
+#> config_file = '/etc/pyromat.py'
 #
 #   also popular is
 #
-#> config_file = '~/.pyro/config.py'
+#> config_file = '~/.pyromat/config.py'
 #
 #   buy you can always get explicit
 #
@@ -69,13 +69,13 @@ config_verbose = False
 # to overwrite previous entries - you can only add additional 
 # directories.
 #
-#> dat_dir = '~/.pyro/data'
+#> dat_dir = '~/.pyromat/data'
 #
 #   or
 #
-#> dat_dir = ['/usr/share/pyro/data', '/home/bestuserever/.pyro/data']
+#> dat_dir = ['/usr/share/pyromat/data', '/home/bestuserever/.pyromat/data']
 
-# Should the PYro dat module print its load activity to stdout?
+# Should the PYroMat dat module print its load activity to stdout?
 #
 dat_verbose = False
 
@@ -102,22 +102,22 @@ dat_recursive = True
 # System administrators should be VERY careful of the latter 
 # implementation.  bestuserever really needs to be an extra special
 # user, because she will be able to write scripts that will be executed
-# as whatever user loads PYro.
+# as whatever user loads PYroMat.
 #
-#> reg_dir = '~/.pyro/reg'
+#> reg_dir = '~/.pyromat/reg'
 #
 #   or
 #
-#> reg_dir = ['/usr/share/pyro/reg', '/home/bestuserever/.pyro/reg']
+#> reg_dir = ['/usr/share/pyromat/reg', '/home/bestuserever/.pyromat/reg']
 
-# Should the pyro.reg.regload() function print its activity to stdout?
+# Should the pyromat.reg.regload() function print its activity to stdout?
 reg_verbose = False
 
 # If redundant class definitions are discovered, should we overwrite the
 # old one, or should we ignore the new one?
 reg_overwrite = True
 
-# Should a redundant class definition cause PYro to exit with an error?
+# Should a redundant class definition cause PYroMat to exit with an error?
 reg_exist_fatal = False
 
 # What is the default temperature and pressure that property functions